Restructure how commands are organized in CLI
