+++ .cargo/config
@@ -1,2 +1,5 @@
 [target.'cfg(not(target_arch = "wasm32"))']
 rustflags = ["-C", "target-cpu=native"]
+
+[build]
+incremental = false # https://github.com/rust-lang/rust/issues/84970
