+++ cli/helpers/mod.rs
@@ -15,4 +15,8 @@
 // along with the Aleo library. If not, see <https://www.gnu.org/licenses/>.
 
 pub mod snarkos;
+
 pub mod transaction;
+
+pub mod updater;
+pub use updater::*;
